#25 [refactor] : JSDoc 및 스타일링 파일 확장자 변경
